fix(DashboardUtils): Add null checks and adjust segment appending
Add null checks to some methods to avoid a `NullPointerException`.
Now an `IllegalArgumentException` is thrown, which is also mentioned in the
method signature, even though it is a `RuntimeException`.
`appendSegmentToPath` is undergoing a complete overhaul and now trims leading/trailing
whitespaces from segments and additional the leading and trailing slash (if there).