test(ServiceInstanceBindingController): Adjust tests for merge code changes
In the last merge from develop, code changes were done to methods which are tested via
unit tests.
Additional `throws Exception` in the method signatures is counterproductive. Declaring every
thrown exception of the method which is tested has the effect, that new declared exceptions
will produce compile-time errors, which then indicates that the tests are not up to date anymore.