test(DashboardUtils): Split up the tests in several subclasses
The previous test class was huge (around 700 LOC).
Now we have a base class and several subclasses extending the base class.
The principle is simple, every `@Nested` annotated class at the first level in the previous big class
is now a subclass of the previous big class. In this case every `@Nested` class was also a
specific method in the class to test.